Created method to replace all favoritas
In loca, it runs a transaction that clears the favorita table and then saves the new favoritas.
In remote, it deletes one-by-one all favoritas and then saves the new ones. This is not used at the moment.

When requesting favoritas the local ones are replaced by the remote ones. This way deleted favoritas get removed.