Adds ObtainUserAction with TDD
It returns an optional of the currently logged in user, if any.
It takes the user from two possible sources: a (singleton) Session or a persisted UserDataSource.
The idea of the Session is to keep a application-wide singleton instance of the current user. Maybe we don't even need it...
Also, when the session is empty and we recover the user from the datasource, we store the user into the session.