Made location optional
LocationProvides now returns optional of location, so we can pass null and get an absent value when the location is not available. Needs some improvement on UX, but at least we are in the same spot as before.

Y had to decompose the subscription in the presenter in two steps, one for the location and another one for the nearby stops.