Migrating firebase to the new sdk
Had to create a different project

Changed the implementation of FavoritaDataSource

Changed LoginAction to write to firebase with the new methods.
To be able to test the LoginAction properly I isolated the communication with FirebaseAuth into a LoginService, which can be easily mocked.
That service is not currently tested. Shame.