Requests new location after GoogleApiClient connected
onResume is not enough, the first and only time it could happend that the google api client is not yet connected and the location won't work.
Now we update the location after the client has been connected. We also do it when navigating through menu items, which is the equivalent to the "onResume" for each fragment.